pyproject.toml
[tool.poetry]
name = "test-project"
version = "0.1.0"
description = ""
authors = ["Nahasapeemapetilon Cetheres <itsjustfortwits@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "2.25.1"
psycopg2 = "^2.9.9"
sqlalchemy = "2.0.22"
pytz = "^2023.3.post1"
python-dotenv = "1.0.1"
openai = "0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
